encounters_deportations folder
Contains all the manually collected datasets from https://deportationdata.org/ that I was unable to use from the apprehensions_origins and removals folders in the data_cleaned folder. Currently put a placeholder text file in order to commit the folder to github. Will add in datasets and upload screenshots of file creation/modification dates of folder and datasets after uploading.